requirements.txt
Flask==2.2.1
keras==2.10.0
numpy==1.23.2
Pillow==9.3.0
tensorflow==2.10.0
pytest==8.3.3
werkzeug==2.3.7
pylint==4.0.1
pre-commit==4.3.0
